Move streaming code to the crate::transcription::live module
