+++ examples/microphone_stream.rs
@@ -92,6 +92,7 @@ async fn main() -> Result<(), DeepgramError> {
     let dg = Deepgram::new(env::var("DEEPGRAM_API_KEY").unwrap());
 
     let mut results = dg
+        .transcription()
         .stream_request()
         .stream(microphone_as_stream())
         // TODO Enum.
